backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.0.1
Werkzeug==3.1.3
openai==1.66.3
python-dotenv==1.0.1
pandas==2.2.3
azure-search-documents==11.4.0b3